fix resolve tab and summary email
